api/requirements.txt
fastapi>=0.109
uvicorn[standard]>=0.27
pandas>=2.1
numpy>=1.26
scikit-learn>=1.4
shap>=0.45
pydantic>=2.5
pydantic-settings>=2.1
openpyxl>=3.1
fairlearn>=0.10
sqlalchemy>=2.0
alembic>=1.13
psycopg2-binary>=2.9
pyjwt>=2.8
bcrypt>=4.1
celery[redis]>=5.3
azure-storage-blob>=12.19
